Start of changes to make the code be able to handle more than 2^32 particles. There is a slight extension to the standard tipsy format which allows very large tipsy files as well. There may still be missing uint64_t type changes here and there.
Also removed UpdateUdot code which was left over from gasoline.

+++ master.h
@@ -28,13 +28,13 @@ typedef struct msrContext {
     ** Other stuff...
     */
     int nThreads;
-    int N;
-    int nDark;
-    int nGas;
-    int nStar;   
-    int nMaxOrder;		/* Order number of last particle */
-    int nMaxOrderGas;
-    int nMaxOrderDark;
+    uint64_t N;
+    uint64_t nDark;
+    uint64_t nGas;
+    uint64_t nStar;   
+    uint64_t nMaxOrder;		/* Order number of last particle */
+    uint64_t nMaxOrderGas;
+    uint64_t nMaxOrderDark;
     int iCurrMaxRung;
     double dCrit;
     /*
